adding communication (correctly parsed)
